+++ Sources/SwiftUIKit/_Deprecated/List/ListHeader.swift
@@ -8,9 +8,7 @@
 
 import SwiftUI
 
-/**
- This view removes the list background and insets for a view.
- */
+@available(*, deprecated, message: "Just use plain list row modifiers instead.")
 @available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
 public struct ListHeader: View {
     
@@ -23,17 +21,6 @@ public struct ListHeader: View {
     public var body: some View {
         Text(text)
             .listRowBackground(Color.clear)
-            .listRowInsets(EdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 0))
-    }
-}
-
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-struct ListHeader_Previews: PreviewProvider {
-    
-    static var previews: some View {
-        List {
-            ListHeader("Text 1")
-            ListText("Text with icon and subtitle")
-        }
+            .listRowInsets(EdgeInsets.all(0))
     }
 }