+++ Sources/SwiftUIKit/_Deprecated/List/ListNavigationButton.swift
@@ -1,0 +1,4 @@
+import SwiftUI
+
+@available(*, deprecated, renamed: "NavigationButton")
+public typealias ListNavigationButton = NavigationButton