+++ Sources/SwiftUIKit/_Deprecated/List/ListText.swift
@@ -8,14 +8,7 @@
 
 import SwiftUI
 
-/**
- This view embeds the provided texts and icon in an `HStack`
- and wraps the stack in a ``ListItem``.
- 
- Use this view when you want to render the content as a full
- list item, with all the padding and other tweaks that comes
- with it. Use ``ListTextContent`` to just get plain content.
- */
+@available(*, deprecated, message: "Just use a plain label instead.")
 @available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
 public struct ListText<Icon: View>: View {
     
@@ -34,12 +27,11 @@ public struct ListText<Icon: View>: View {
     private let subtitle: String?
     
     public var body: some View {
-        ListItem {
-            ListTextContent(text, icon: icon, subtitle: subtitle)
-        }
+        ListTextContent(text, icon: icon, subtitle: subtitle)
     }
 }
 
+@available(*, deprecated, message: "Just use a plain label instead.")
 @available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
 public extension ListText where Icon == EmptyView {
     
@@ -53,6 +45,7 @@ public extension ListText where Icon == EmptyView {
     }
 }
 
+@available(*, deprecated, message: "Just use a plain label instead.")
 @available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
 public extension ListText where Icon == Image {
     
@@ -64,26 +57,5 @@ public extension ListText where Icon == Image {
         self.text = text
         self.icon = icon
         self.subtitle = subtitle
-    }
-}
-
-@available(iOS 14.0, macOS 11.0, tvOS 14.0, watchOS 7.0, *)
-struct ListText_Previews: PreviewProvider {
-    
-    static var previews: some View {
-        List {
-            ListText("Text 1")
-            ListText("Text 2")
-            ListText(
-                "Text with icon",
-                icon: Image(systemName: "lightbulb"))
-            ListText(
-                "Text with color badge",
-                icon: Color.blue.clipShape(Circle()).padding(1))
-            ListText(
-                "Text with icon and subtitle",
-                icon: Image(systemName: "lightbulb"),
-                subtitle: "Subtitle")
-        }
     }
 }